Prevent scroll on Reload/Refresh
Chrome/Edge appears to loose the location fragment focus on Refresh, so
I'm focusing the fragment target explictly after loading the page, at
least if it is a room.  However, this causes Firefox to "loose" its saved
scroll position on Reload, so prevent scrolling to the selected room
during that time.

+   Update lastScroll on "pageshow" to ensure up-to-date in the abscense
    of "scroll" events during page load